2003-07-27  Dalibor Topic <robilad@kaffe.org>,             Helmer Kraemer <hkraemer@freenet.de>
        Fixes for compiler warnings on alpha-linux.

        * config/alpha/alpha.c:
        (__alpha_ra) use uintp instead of int*.
        (__alpha_nextFrame) use uintp instead of char*.

        * config/alpha/linux/jit-md.h:
        (_exceptionFrame)  use uintp instead of char*.
        (__alpha_ra) use uintp instead of int*.

        * kaffe/kaffevm/classMethod.h:
        (methods.idx) changed type to long.

        * kaffe/kaffevm/jni.c:
        (KAFFE_JNI_SETEXCEPTFP) added a cast to uintp.

        * kaffe/kaffevm/jit/icode.c:
        (softcall_lookupinterfacemethod) removed unnecessary cast to int.

+++ kaffe/kaffevm/classMethod.h
@@ -241,7 +241,7 @@ typedef struct _methods {
 	Utf8Const*		name;
 	parsed_signature_t*	parsed_sig;
 	accessFlags		accflags;
-	short			idx;	/* Index into class->dtable */
+	long			idx;	/* Index into class->dtable */
 	u2			stacksz;
 	u2			localsz;
 	/* Only used for static/final/constructor methods */
